Moved _populate_normalize_config to sql/operator/load_file.py
